Lots of changes. Read further.
Small formatting changes (whitespace and stuff)

Added support for slowly changing parameters.

All the slowly changin parameters are now
cycled in a function. This means that for
each sensor scan only one of the slow parameters
is sent each time. This effectively makes
the refresh rate of each slow parameter
inverselly proportional to the number of
slow parameters.

Refactored MavlinkProcessor::MavlinkTelemetry
to a struct which describes general telemetry.
The name of the struct is Telemetry and was
moved to FrSkySmartPort.h

Most of the enums in FrSkyProcessor.h
are moved to a FrSkySmartPort.h header.
This enables a single header to have all
the information necessary to work with
FrSkySmartPort telemetry.

Added a number of new telemetry that
the converter sends. Sepecifically added:
system status, mav type, roll, yaw,
pitch and complete base_mode. This additions
use FrSkyIDs which are not yet attributed and
they may break in future OpenTX releases.

Heartbeat base modes are now sent according
to the real heartbeats received. This way
there is an implicit way to know if the drone
is operational in the mavlink sense.

Added new modes in Lua but some are still
missing.